+++ projects/plugins/boost/app/class-jetpack-boost.php
@@ -17,6 +17,7 @@ use Automattic\Jetpack_Boost\Admin\Admin;
 use Automattic\Jetpack_Boost\Admin\Config;
 use Automattic\Jetpack_Boost\Admin\Regenerate_Admin_Notice;
 use Automattic\Jetpack_Boost\Features\Optimizations\Optimizations;
+use Automattic\Jetpack_Boost\Features\Setup_Prompt\Setup_Prompt;
 use Automattic\Jetpack_Boost\Lib\Analytics;
 use Automattic\Jetpack_Boost\Lib\CLI;
 use Automattic\Jetpack_Boost\Lib\Connection;
@@ -98,6 +99,10 @@ class Jetpack_Boost {
 
 		// Initialize the Admin experience.
 		$this->init_admin( $optimizations );
+
+		// Add the setup prompt.
+		Setup::add( new Setup_Prompt() );
+
 		add_action( 'init', array( $this, 'init_textdomain' ) );
 
 		add_action( 'handle_environment_change', array( $this, 'handle_environment_change' ) );
